requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.30.0
python-dotenv>=1.0.0
notebook>=6.5.0
